okod upgrade

+++ src/Service/RankingScoreService.php
@@ -118,7 +118,7 @@ class RankingScoreService
             'plateform' => $isVr ? 'vr' : ($isOkod ? 'flat_okod' : 'flat'),
         ]);
 
-        if ($totalPondPPScore == 0) {
+        if ($totalPondPPScore == 0 && $rankedScore) {
             $this->rankedScoresRepository->remove($rankedScore);
 
             return;
